// src/client/zod.gen.ts
export const zFamilyMemberCreate = z.object({
    relationship: z.string(),
    name: z.string(),
    birth_date: z.optional(z.union([
        z.iso.date(),
        z.null()
    ])),
    blessed: z.optional(z.boolean()).default(false),
    remarks: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    display_order: z.optional(z.int()).default(0),
    resume_id: z.optional(z.union([
        z.uuid(),
        z.null()
    ]))
});